Add: REST para ClientePF

+++ src/main/java/br/gov/sp/fatec/lab5/controller/ClientePFController.java
@@ -2,42 +2,51 @@ package br.gov.sp.fatec.lab5.controller;
 
 import br.gov.sp.fatec.lab5.entity.Cliente;
 import br.gov.sp.fatec.lab5.entity.ClientePF;
+import br.gov.sp.fatec.lab5.entity.Usuario;
+import br.gov.sp.fatec.lab5.entity.View;
 import br.gov.sp.fatec.lab5.service.ClienteService;
+import com.fasterxml.jackson.annotation.JsonView;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.util.UriComponentsBuilder;
 
 @CrossOrigin
 @RestController
-@RequestMapping("/clientesPF")
+@RequestMapping("/clientePF")
 public class ClientePFController {
 
     @Autowired
     private ClienteService service;
 
+    @JsonView(View.ClienteSimplesPF.class)
     @GetMapping
     public ResponseEntity<Iterable<Cliente>> buscarTodos(){
         return ResponseEntity.ok(service.findAll());
     }
 
+    @JsonView(View.ClienteCompletoPF.class)
     @GetMapping("/{id}")
     public ResponseEntity<Cliente> buscarPorId(@PathVariable Long id){
-        //TODO: retornar 404 caso o id não exista
-        //TODO: tratar erro caso id == null
-        return ResponseEntity.ok(service.buscarPorId(id));
+        Cliente cliente = service.buscarPorId(id);
+        if (cliente == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(cliente);
     }
 
+    @JsonView(View.ClienteSimplesPF.class)
     @GetMapping("/q")
-    public ResponseEntity<Cliente> findByNome(@RequestParam String nome){
+    public ResponseEntity<Iterable<Cliente>> findByNome(@RequestParam String nome){
         return ResponseEntity.ok(service.buscarPorNome(nome));
     }
 
     @PostMapping
-    public ResponseEntity salvar(@RequestBody ClientePF cliente) {
+    public ResponseEntity salvar(@RequestBody ClientePF cliente, UriComponentsBuilder uriComponentsBuilder) {
         cliente.setId(null);
         service.save(cliente);
-        //TODO: alterar para retornar a URL com id do recurso criado
-        return ResponseEntity.ok(cliente);
+        return ResponseEntity.created(uriComponentsBuilder.path(
+                "/clientePF/" + cliente.getId()).build().toUri()).build();
     }
 
     @PutMapping
@@ -47,8 +56,10 @@ public class ClientePFController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity deletar(@PathVariable Long id){
-
-        // TODO: retornar 404 se o item a ser deletado não existir.
+        Cliente cliente = service.buscarPorId(id);
+        if (cliente == null) {
+            return ResponseEntity.notFound().build();
+        }
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
